Make branch compatible to run e2e tests on local.
