+++ src/e2e-test/java/io/cdap/plugin/salesforcebatchsource/actions/SalesforcePropertiesPageActions.java
@@ -17,10 +17,8 @@
 package io.cdap.plugin.salesforcebatchsource.actions;
 
 import io.cdap.cdap.api.data.schema.Schema;
-import io.cdap.e2e.utils.AssertionHelper;
 import io.cdap.e2e.utils.PluginPropertyUtils;
 import io.cdap.e2e.utils.SeleniumHelper;
-import io.cdap.e2e.utils.WaitHelper;
 import io.cdap.plugin.salesforce.SalesforceConnectionUtil;
 import io.cdap.plugin.salesforce.authenticator.Authenticator;
 import io.cdap.plugin.salesforce.authenticator.AuthenticatorCredentials;
@@ -34,9 +32,7 @@ import io.cdap.plugin.utils.SchemaFieldTypeMapping;
 import io.cdap.plugin.utils.SchemaTable;
 import io.cdap.plugin.utils.enums.SOQLQueryType;
 import io.cdap.plugin.utils.enums.SObjects;
-import io.cdap.plugin.utils.enums.SalesforceBatchSourceProperty;
 import org.apache.commons.lang3.RandomStringUtils;
-import org.openqa.selenium.WebElement;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -54,7 +50,7 @@ public class SalesforcePropertiesPageActions {
 
   public static void fillReferenceName(String referenceName) {
     logger.info("Fill Reference name: " + referenceName);
-    SalesforcePropertiesPage.referenceInput.sendKeys(referenceName);
+    SalesforcePropertiesPage.referenceNameInput.sendKeys(referenceName);
   }
 
   public static void fillAuthenticationProperties(String username, String password, String securityToken,
@@ -69,11 +65,11 @@ public class SalesforcePropertiesPageActions {
 
   public static void fillAuthenticationPropertiesForSalesforceAdminUser() {
     SalesforcePropertiesPageActions.fillAuthenticationProperties(
-      PluginPropertyUtils.pluginProp("admin.username"),
-      PluginPropertyUtils.pluginProp("admin.password"),
-      PluginPropertyUtils.pluginProp("admin.security.token"),
-      PluginPropertyUtils.pluginProp("admin.consumer.key"),
-      PluginPropertyUtils.pluginProp("admin.consumer.secret"));
+      System.getenv("SALESFORCE_USERNAME"),
+      System.getenv("SALESFORCE_PASSWORD"),
+      System.getenv("SALESFORCE_SECURITY_TOKEN"),
+      System.getenv("SALESFORCE_CONSUMER_KEY"),
+      System.getenv("SALESFORCE_CONSUMER_SECRET"));
   }
 
   public static void fillAuthenticationPropertiesWithInvalidValues() {
@@ -105,27 +101,6 @@ public class SalesforcePropertiesPageActions {
     fillReferenceName(referenceName);
     fillSObjectName(sObjectName.value);
   }
-
-  public static void clickOnGetSchemaButton() {
-    logger.info("Click on the Get Schema button");
-    SalesforcePropertiesPage.getSchemaButton.click();
-    WaitHelper.waitForElementToBeDisplayed(SalesforcePropertiesPage.loadingSpinnerOnGetSchemaButton);
-    WaitHelper.waitForElementToBeHidden(SalesforcePropertiesPage.loadingSpinnerOnGetSchemaButton);
-    WaitHelper.waitForElementToBeDisplayed(SalesforcePropertiesPage.getSchemaButton);
-  }
-
-  public static void clickOnValidateButton() {
-    logger.info("Click on the Validate button");
-    SalesforcePropertiesPage.validateButton.click();
-    WaitHelper.waitForElementToBeDisplayed(SalesforcePropertiesPage.loadingSpinnerOnValidateButton);
-    WaitHelper.waitForElementToBeHidden(SalesforcePropertiesPage.loadingSpinnerOnValidateButton);
-    WaitHelper.waitForElementToBeDisplayed(SalesforcePropertiesPage.validateButton);
-  }
-
-  public static void verifyNoErrorsFoundSuccessMessage() {
-    AssertionHelper.verifyElementDisplayed(SalesforcePropertiesPage.noErrorsFoundSuccessMessage);
-  }
-
   private static AuthenticatorCredentials setAuthenticationCredentialsOfAdminUser() {
     return new AuthenticatorCredentials(PluginPropertyUtils.pluginProp("admin.username"),
       PluginPropertyUtils.pluginProp("admin.password"),
@@ -220,65 +195,6 @@ public class SalesforcePropertiesPageActions {
     return expectedSchemaTable;
   }
 
-  public static void verifyFieldTypeMappingDisplayed(SchemaFieldTypeMapping schemaFieldTypeMapping) {
-    WebElement element = SalesforcePropertiesPage.getSchemaFieldTypeMappingElement(schemaFieldTypeMapping);
-    AssertionHelper.verifyElementDisplayed(element);
-  }
-
-  public static void verifyOutputSchemaTable(SchemaTable schemaTable) {
-    List<SchemaFieldTypeMapping> listOfFields = schemaTable.getListOfFields();
-
-    for (SchemaFieldTypeMapping schemaFieldTypeMapping : listOfFields) {
-      verifyFieldTypeMappingDisplayed(schemaFieldTypeMapping);
-    }
-  }
-
-  public static void clickOnClosePropertiesPageButton() {
-    logger.info("Close the Salesforce (batch) source properties page");
-    SalesforcePropertiesPage.closePropertiesPageButton.click();
-  }
-
-  public static void verifyRequiredFieldsMissingValidationMessage(SalesforceBatchSourceProperty propertyName) {
-    WebElement element = SalesforcePropertiesPage.getPropertyInlineErrorMessage(propertyName);
-
-    AssertionHelper.verifyElementDisplayed(element);
-    AssertionHelper.verifyElementContainsText(element, propertyName.propertyMissingValidationMessage);
-  }
-
-  public static void verifyPropertyInlineErrorMessage(SalesforceBatchSourceProperty property,
-                                                      String expectedErrorMessage) {
-    WebElement element = SalesforcePropertiesPage.getPropertyInlineErrorMessage(property);
-
-    AssertionHelper.verifyElementDisplayed(element);
-    AssertionHelper.verifyElementContainsText(element, expectedErrorMessage);
-  }
-
-  public static void verifyInvalidSoqlQueryErrorMessageForStarQueries() {
-    verifyPropertyInlineErrorMessage(SalesforceBatchSourceProperty.SOQL_QUERY,
-      PluginPropertyUtils.errorProp("invalid.soql.starquery"));
-  }
-
-  public static void verifyErrorMessageOnHeader(String expectedErrorMessage) {
-    AssertionHelper.verifyElementContainsText(SalesforcePropertiesPage.errorMessageOnHeader,
-      expectedErrorMessage);
-  }
-
-  public static void verifyValidationMessageForBlankAuthenticationProperty() {
-    verifyErrorMessageOnHeader(PluginPropertyUtils.errorProp("empty.authentication.property"));
-  }
-
-  public static void verifyValidationMessageForInvalidAuthenticationProperty() {
-    verifyErrorMessageOnHeader(PluginPropertyUtils.errorProp("invalid.authentication.property"));
-  }
-
-  public static void verifyValidationMessageForMissingSoqlOrSobjectNameProperty() {
-    verifyErrorMessageOnHeader(PluginPropertyUtils.errorProp("required.property.soqlorsobjectname.error"));
-    verifyRequiredFieldsMissingValidationMessage(SalesforceBatchSourceProperty.SOQL_QUERY);
-  }
-
-  public static void verifyValidationMessageForInvalidSObjectName(String invalidSObjectName) {
-    String expectedValidationMessage = PluginPropertyUtils.errorProp(
-      "invalid.sobjectname.error") + " '" + invalidSObjectName + "''";
-    verifyErrorMessageOnHeader(expectedValidationMessage);
+  public static void verifyIfRecordCreatedInSinkForObjectIsCorrect(String expectedOutputFile) {
   }
 }
